+++ src/layouts/InterfaceLayout/containers/SignMessageContainer/SignMessageContainer.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="deploy-contract-container">
-    <success-modal
-      message=""
+    <success-modal 
+      message="" 
       link-message="Ok"/>
     <interface-container-title :title="$t('common.signMessage')"/>
     <div class="send-form">
@@ -22,9 +22,7 @@
       </div>
 
       <div class="the-form">
-        <textarea
-          ref="message"
-          class="custom-textarea-1"/>
+        <textarea class="custom-textarea-1"/>
       </div>
     </div>
 
@@ -35,31 +33,31 @@
           <popover :popcontent="$t('popover.whatIsSignatureContent')"/>
 
           <div class="copy-buttons">
-            <span @click="deleteInputText('signature')">Clear</span>
-            <span @click="copyToClipboard('signature')">Copy</span>
+            <span @click="deleteInputText('abi')">Clear</span>
+            <span @click="copyToClipboard('abi')">Copy</span>
           </div>
 
         </div>
       </div>
       <div class="the-form domain-name">
-        <textarea
-          ref="signature"
-          class="custom-textarea-1"
+        <textarea 
+          ref="abi" 
+          class="custom-textarea-1" 
           name=""/>
       </div>
     </div>
 
     <div class="submit-button-container">
       <div class="buttons">
-        <div
-          class="submit-button large-round-button-green-filled clickable"
-          @click="signMessage">
+        <div 
+          class="submit-button large-round-button-green-filled clickable" 
+          @click="successModalOpen">
           {{ $t('Sign') }}
         </div>
       </div>
-      <interface-bottom-text
-        :link-text="$t('interface.learnMore')"
-        :question="$t('interface.haveIssues')"
+      <interface-bottom-text 
+        :link-text="$t('interface.learnMore')" 
+        :question="$t('interface.haveIssues')" 
         link="/"/>
     </div>
 
@@ -69,7 +67,7 @@
 <script>
 import InterfaceBottomText from '@/components/InterfaceBottomText';
 import InterfaceContainerTitle from '../../components/InterfaceContainerTitle';
-import SuccessModal from '@/containers/ConfirmationContainer/components/SuccessModal/SuccessModal.vue';
+import SuccessModal from '@/components/SuccessModal';
 
 export default {
   name: 'SignMessage',
@@ -82,37 +80,12 @@ export default {
     return {};
   },
   methods: {
-    signMessage() {
-      this.$store.state.web3.eth
-        .sign(
-          this.$refs.message.value,
-          this.$store.state.wallet.getAddressString()
-        )
-        .then(_signedMessage => {
-          this.$refs.signature.value = JSON.stringify(
-            {
-              address: this.$store.state.wallet.getAddressString(),
-              msg: this.$refs.message.value,
-              sig: _signedMessage,
-              version: '3',
-              signer: this.$store.state.wallet.brand
-                ? this.$store.state.wallet.brand
-                : 'MEW'
-            },
-            null,
-            2
-          );
-        })
-        // eslint-disable-next-line
-        .catch(console.error);
-    },
     successModalOpen() {
       this.$children[0].$refs.success.show();
     },
     copyToClipboard(ref) {
       this.$refs[ref].select();
       document.execCommand('copy');
-      window.getSelection().removeAllRanges();
     },
     deleteInputText(ref) {
       this.$refs[ref].value = '';