+++ src/layouts/InterfaceLayout/containers/SwapContainer/SwapContainer.vue
@@ -1,10 +1,12 @@
 <template>
   <div class="swap-container">
+    <swap-confirmation-modal></swap-confirmation-modal>
+
     <div class="title-block">
       <interface-container-title :title="$t('common.swap')"/>
-      <div class="images">
-        <img :src="images.kybernetowrk">
-        <img :src="images.bity">
+      <div class="buy-eth">
+        <span>Buy ETH with</span>
+        <img :src="images.visaMaster">
       </div>
     </div>
 
@@ -27,9 +29,7 @@
           </div>
         </div>
         <div class="exchange-icon">
-          <i
-            class="fa fa-exchange"
-            aria-hidden="true"/>
+          <img :src="images.swap">
         </div>
         <div class="amount">
           <div class="title">
@@ -52,20 +52,26 @@
 
     <div class="send-form">
       <div class="title-container">
-        <div class="title">
+        <div class="title title-and-copy">
           <h4>{{ $t('common.toAddress') }}</h4>
+          <p class="copy-button prevent-user-select">Copy</p>
         </div>
       </div>
       <div class="the-form gas-amount">
-        <input
-          type="number"
-          name=""
-          value=""
-          placeholder="Please Enter The Address" >
+        <drop-down-address-selector></drop-down-address-selector>
       </div>
     </div>
 
     <div class="send-form">
+      <div class="title-container">
+        <div class="title title-and-copy">
+          <h4>Providers</h4>
+        </div>
+      </div>
+      <providers-radio-selector></providers-radio-selector>
+    </div>
+
+    <div class="send-form" v-if="false">
       <div class="title-container">
         <div class="title">
           <div class="title-and-popover">
@@ -103,42 +109,45 @@
     </div>
 
     <div class="submit-button-container">
-      <h4>1 ETH = 0.000231 BTC</h4>
-      <div class="submit-button large-round-button-green-filled clickable">
+      <h4 v-if="false">1 ETH = 0.000231 BTC</h4>
+      <div class="submit-button large-round-button-green-filled clickable" v-on:click="swapConfirmationModalOpen">
         {{ $t('common.continue') }}
         <i
           class="fa fa-long-arrow-right"
           aria-hidden="true"/>
-      </div>
-      <div class="buy-eth">
-        <span>Buy ETH with</span>
-        <img :src="images.visaMaster">
       </div>
     </div>
 
   </div>
 </template>
 <script>
-import CurrencyPicker from '../../components/CurrencyPicker';
-import InterfaceBottomText from '@/components/InterfaceBottomText';
-import InterfaceContainerTitle from '../../components/InterfaceContainerTitle';
-
-import ImageKybernetowrk from '@/assets/images/etc/kybernetowrk.png';
-import ImageBity from '@/assets/images/etc/bity.png';
-import ImageVisaMaster from '@/assets/images/etc/visamaster.png';
+import ProvidersRadioSelector from '../../components/ProvidersRadioSelector'
+import CurrencyPicker from '../../components/CurrencyPicker'
+import DropDownAddressSelector from '@/components/DropDownAddressSelector'
+import InterfaceBottomText from '@/components/InterfaceBottomText'
+import InterfaceContainerTitle from '../../components/InterfaceContainerTitle'
+import swapIcon from '@/assets/images/icons/swap.svg'
+import ImageKybernetowrk from '@/assets/images/etc/kybernetowrk.png'
+import ImageBity from '@/assets/images/etc/bity.png'
+import ImageVisaMaster from '@/assets/images/etc/visamaster.png'
+import SwapConfirmationModal from './components/SwapConfirmationModal'
 
 export default {
   components: {
     'interface-bottom-text': InterfaceBottomText,
     'interface-container-title': InterfaceContainerTitle,
-    'currency-picker': CurrencyPicker
+    'currency-picker': CurrencyPicker,
+    'drop-down-address-selector': DropDownAddressSelector,
+    'providers-radio-selector': ProvidersRadioSelector,
+    'swap-confirmation-modal': SwapConfirmationModal
   },
   data() {
     return {
       images: {
         kybernetowrk: ImageKybernetowrk,
         bity: ImageBity,
-        visaMaster: ImageVisaMaster
+        visaMaster: ImageVisaMaster,
+        swap: swapIcon
       },
       toArray: [
         { symbol: 'BTC', name: 'Bitcoin' },
@@ -153,8 +162,8 @@ export default {
     };
   },
   methods: {
-    copyAddress() {
-      alert("This doesn't work for now.");
+    swapConfirmationModalOpen () {
+      this.$children[0].$refs.swapconfirmation.show()
     }
   }
 };