+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Let users deploy and/or interact with contracts [#45](https://github.com/MyEtherWallet/MyEtherWallet/pull/45)
 
 ### UI/UX
+- Add Swap confirmation modal [#78](https://github.com/MyEtherWallet/MyEtherWallet/pull/78)
+- Add FAQs page [#58](https://github.com/MyEtherWallet/MyEtherWallet/pull/58)
 - Add Create Wallet warning page [#46](https://github.com/MyEtherWallet/MyEtherWallet/pull/46)
 
 ### Devop